+++ src/event.c
@@ -228,13 +228,12 @@ update_ewmh_active_win(struct x_connection * /*c*/, struct x_async_request_base 
 	auto wid = *(xcb_window_t *)xcb_get_property_value(reply);
 	log_debug("EWMH _NET_ACTIVE_WINDOW is %#010x", wid);
 
-	auto cursor = wm_find_by_client(ps->wm, wid);
-	auto w = cursor ? wm_ref_deref(cursor) : NULL;
-
 	// Mark the window focused. No need to unfocus the previous one.
-	if (w) {
-		win_set_focused(ps, w);
-	}
+	auto cursor = wm_find_by_client(ps->wm, wid);
+	wm_ref_set_focused(ps->wm, cursor);
+	ps->pending_updates = true;
+	log_debug("%#010x (%s) focused.", wid,
+	          wm_ref_deref(cursor) ? wm_ref_deref(cursor)->name : "");
 }
 
 struct ev_recheck_focus_request {
@@ -295,11 +294,10 @@ static void recheck_focus(struct x_connection * /*c*/, struct x_async_request_ba
 	}
 
 	// And we set the focus state here
-	auto w = wm_ref_deref(cursor);
-	if (w) {
-		log_debug("%#010x (%s) focused.", wid, w->name);
-		win_set_focused(ps, w);
-	}
+	wm_ref_set_focused(ps->wm, cursor);
+	ps->pending_updates = true;
+	log_debug("%#010x (%s) focused.", wid,
+	          wm_ref_deref(cursor) ? wm_ref_deref(cursor)->name : "");
 }
 
 void ev_update_focused(struct session *ps) {
@@ -388,6 +386,7 @@ static void configure_win(session_t *ps, xcb_configure_notify_event_t *ce) {
 
 	auto w = wm_ref_deref(cursor);
 	if (!w) {
+		log_debug("Window %#010x is unmanaged.", ce->window);
 		return;
 	}